% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsCalculatePriorityScores.R
\name{CalculatePriorityScores}
\alias{CalculatePriorityScores}
\title{Calculate priority scores}
\usage{
CalculatePriorityScores(result, useRNAvaf = FALSE)
}
\arguments{
\item{result}{A results file generated from main functions}

\item{useRNAvaf}{To indicate whether this function uses DNA VAF or RNA VAF for the calculation.}
}
\value{
P_I: Priority score using the IC50.

P_R: Priority score using the percentage of rank affinity.

P: Priority score implemented in MuPeXI (Bjerregaard et al. 2017).
}
\description{
Calculate priority scores
}
