% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryFragments.R
\name{Export_Summary_Fragments}
\alias{Export_Summary_Fragments}
\title{Export Summary Count from Indel/SV Results}
\usage{
Export_Summary_Fragments(Input, mut_IC50_th, Total_RNA_th = NA,
  Tumor_RNA_th = NA, MutRatio_th = NA, WriteLongIndel = NA)
}
\arguments{
\item{Input}{Input file generated from MainSNVClass1,2.}

\item{mut_IC50_th}{The threshold for mutant peptide to be neoantigen.}

\item{Total_RNA_th}{The total RNA expression threshold.}

\item{Tumor_RNA_th}{The tumor specific RNA expression threshold.}

\item{MutRatio_th}{The mutation ratio threshold.}

\item{WriteLongIndel}{If setting a file name, Write Long Indels of which the p-value is less than 0.05.}
}
\value{
Num_Alteration The number of evaluated alterations.

Num_Alteration_Generating_NeoAg The number of evaluated alterations that can generate neoantigen.

Num_Peptide The number of evaluated peptifdes.

Num_Peptide_Generating_NeoAg The number of evaluated peptides that can be neoantigen.
}
\description{
Export Summary Count from Indel/SV Results
}
