% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarySNV.R
\name{Export_Summary_SNV}
\alias{Export_Summary_SNV}
\title{Export Summary Count from SNV Results}
\usage{
Export_Summary_SNV(Input, mut_IC50_th, wt_IC50_th = NA, Total_RNA_th = NA,
  Tumor_RNA_th = NA, MutRatio_th = NA)
}
\arguments{
\item{Input}{Input file generated from MainSNVClass1,2.}

\item{mut_IC50_th}{The threshold for mutant peptide to be neoantigen.}

\item{wt_IC50_th}{The threshold for wt peptide to be neoantigen.}

\item{Total_RNA_th}{The total RNA expression threshold.}

\item{Tumor_RNA_th}{The tumor specific RNA expression threshold.}

\item{MutRatio_th}{The mutation ratio threshold.}
}
\value{
Num_Alteration The number of evaluated alterations.

Num_Alteration_Generating_NeoAg The number of evaluated alterations that can generate neoantigen.

Num_Peptide The number of evaluated peptifdes.

Num_Peptide_Generating_NeoAg The number of evaluated peptides that can be neoantigen.
}
\description{
Export Summary Count from SNV Results
}
