% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{hla_table2}
\alias{hla_table2}
\title{Prices of 50,000 round cut diamonds}
\format{A data frame with 53940 rows and 10 variables}
\usage{
data(hla_table2)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item price. price in US dollars (\$326--\$18,823)
}
}
\keyword{datasets}
