% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallRefFlat.R
\name{InstallRefFlat}
\alias{InstallRefFlat}
\title{Get refFlat file}
\usage{

  InstallRefFlat(url = "http://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/refFlat.txt.gz",
  export_dir = "lib")
}
\arguments{
\item{url}{Url for getting the corresponding refFlat.txt.gz
aaa
(Default = "http://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/refFlat.txt.gz").}

\item{export_dir}{Export directory (Default = "lib").}
}
\value{
void
}
\description{
Get refFlat file
}
