% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallRefMrnaFile.R
\name{InstallRefMrnaFile}
\alias{InstallRefMrnaFile}
\title{Get refMrna file}
\usage{
InstallRefMrnaFile(url = NA, export_dir = "lib")
}
\arguments{
\item{url}{Url for getting the corresponding refMrna.fa.gz

(Default = "http://hgdownload.cse.ucsc.edu/goldenPath/hg19/bigZips/refMrna.fa.gz").}

\item{export_dir}{Export directory (Default = "lib").}
}
\value{
void
}
\description{
Get refMrna file
}
