% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSam.R
\name{InstallSam}
\alias{InstallSam}
\title{Install Samtools}
\usage{
InstallSam(export_dir = "lib")
}
\arguments{
\item{export_dir}{Export directory for using Sample Files (Default="lib").}
}
\value{
void
}
\description{
Install Samtools
}
