% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSampleFiles.R
\name{InstallSampleFiles}
\alias{InstallSampleFiles}
\title{Get Sample Files for Neoantimon}
\usage{
InstallSampleFiles(export_dir = "lib")
}
\arguments{
\item{export_dir}{Export directory for using Sample Files (Default="lib").}
}
\value{
void
}
\description{
Get Sample Files for Neoantimon
}
