% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSamtools.R
\name{InstallSamtools}
\alias{InstallSamtools}
\title{Install Samtools}
\usage{

  InstallSamtools(url = "https://github.com/hase62/Neoantimon/raw/master/lib/samtools-0.1.19.tar.bz2",
  export_dir = "lib")
}
\arguments{
\item{url}{Url for getting samtools
(Default = "https://github.com/hase62/Neoantimon/raw/master/lib/samtools-0.1.19.tar.bz2").}

\item{export_dir}{Export directory (Default = "lib").}
}
\value{
void
}
\description{
Install Samtools
}
