% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainEntireRegionClass2.R
\name{MainEntireRegionClass2}
\alias{MainEntireRegionClass2}
\title{Calculate A Set All Neoantigen Candidates from A Given Gene Symbol and nm_id for MHC Class2 (Not yet stably available)}
\usage{
MainEntireRegionClass2(
  input_nm_id,
  group_ids = seq(1:length(input_nm_id)),
  hla_file = "here_is_a_table",
  hla_types = NA,
  file_name_in_hla_table = NA,
  refflat_file = paste(hmdir, "lib/refFlat.txt", sep = "/"),
  refmrna_file = paste(hmdir, "lib/refMrna.fa", sep = "/"),
  hmdir = getwd(),
  job_id = "ID",
  export_dir = paste("result", job_id, "EntireRegion2", sep = "."),
  netMHCIIpan_dir = paste(hmdir, "lib/netMHCIIpan-3.1/netMHCIIpan", sep = "/"),
  peptide_length = c(15),
  reading_frame = 1,
  CalculateEL = FALSE,
  ignore_short = TRUE
)
}
\arguments{
\item{input_nm_id}{(Required) An input amino acid sequence indicated as NM_ID}

\item{group_ids}{flag to cluster the same group}

\item{hla_file}{A tab separated file indicating HLA types.
The 1st column is input_file name, and the following columns indicate HLA types.

See by data(sample_hla_table_c1); sample_hla_table_c1;}

\item{hla_types}{Set a list of HLA types}

\item{file_name_in_hla_table}{If the name (1st column) in HLA table is not the same as input_file, indicate the corresponding name (Default=input_file).}

\item{refflat_file}{refFlat file to be used in constructing peptide. (Default=paste(hmdir, "lib/refFlat.txt", sep="").

See "https://github.com/hase62/Neoantimon"}

\item{refmrna_file}{refMrna file to be used in constructing peptide (Default=paste(hmdir, "lib/refMrna.fa", sep="").

See "https://github.com/hase62/Neoantimon"}

\item{hmdir}{Home directory for the analysis (Default = getwd()).}

\item{job_id}{Job-id to be attached in output files (Default = "NO_job_id").}

\item{export_dir}{The directory will be stored results (Default = "paste("result", file_name_in_hla_table, job_id, sep=".")")}

\item{netMHCIIpan_dir}{The file directory to netMHCpan (Default="lib/netMHCIIpan-3.2/netMHCIIpan").}

\item{peptide_length}{Peptide Length to be generated (Default = {8,9,10,11,12,13}).}

\item{reading_frame}{The starting frame of the input sequence (Default = 1)}

\item{CalculateEL}{Whether Calculate EL by NetMHCpan or not.}

\item{ignore_short}{Ignore to output results of Short Peptide Less Than min(peptide_length)}
}
\value{
void (Calculated Neoantigen Files will be generated as .tsv files.):

HLA:  HLA type used to calculate neoantigen.

Pos:  The position of a fraction of peptide used to be evaluated from the full-length peptide.

Gene:  Gene symbol used to be evaluated in NetMHCpan.

Evaluated_Mutant_Peptide:  The mutant peptide to be evaluated.

Evaluated_Mutant_Peptide_Core:  The core peptide of the mutant peptide to be evaluated in NetMHCpan.

Mut_EL: EL value for evaluated mutant peptide.

Mut_Rank: Rank value for evaluated mutanat peptide.

Chr: Chromosome Number of the mutation.

NM_ID: NM_ID used to construct peptides from the mutation.

Change: The annotation to be described in .vcf file.

Ref: reference type nucleic acid base.

Alt: alternative type nucleic acid base.

Prob: A probability of reference nucleic acid base described in .vcf file.

Mutation_Prob: A probability of alternative nucleic acid base described in .vcf file.

Exon_Start: The exon start position of the corrsponding NM_ID.

Exon_End: The exon end position of the corrsponding NM_ID.

Mutation_Position: The mutation position of the corrsponding NM_ID.

Total_Depth: The sum depth of the reference and alternative nucleic acid base.

Tumor_Depth: The depth of the alternative nucleic acid base.

Wt_Peptide: The full-length of the wild-type peptide.

Mutant_Peptide: The full-length of the mutant peptide.

Total_RNA: The expression amount of the corresponding RNA.

Tumor_RNA_Ratio: The variant allele frequency of the corresponding RNA.

Tumor_RNA: The modified amount of the corresponding RNA level based on RNA Reads.

Tumor_RNA_based_on_DNA: The modified amount of the corresponding RNA level based on DNA Reads.

MutRatio: The mean value of the cancer cell fraction probability.

MutRatio_Min: The 1\% percentile of the cancer cell fraction probability.

MutRatio_Max: The 99\% percentile of the cancer cell fraction probability.
}
\description{
Calculate A Set All Neoantigen Candidates from A Given Gene Symbol and nm_id for MHC Class2 (Not yet stably available)
}
