% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainMergeSNVClass1.R
\name{MainMergeClass1}
\alias{MainMergeClass1}
\title{Merge Results from MainSnvClass1.R}
\usage{
MainMergeClass1(hmdir = getwd(), input_dir, input_file_prefix,
  Tumor_RNA_BASED_ON_DNA = TRUE, INDEL = FALSE)
}
\arguments{
\item{hmdir}{Home directory for the analysis (Default=getwd()).}

\item{input_dir}{Directory storing netMHCpan Results (Required).}

\item{input_file_prefix}{File prefix of netMHCpan Results (Required).
If you have "sample_annovar.txt.NO_JOB_ID.HLACLASS1.1.peptide.txt", please set "sample_annovar", "sample_annovar.txt" or "sample_annovar.txt.NO_JOB_ID".}

\item{Tumor_RNA_BASED_ON_DNA}{In calculating tumor specific RNA expression, TRUE uses variant allele frequency on DNA. Otherwise, use VAF on RNA (Default=TRUE).}

\item{INDEL}{If the targeting results are generated from Indels, Please check TRUE.}
}
\value{
void (Calculated Neoantigen Files will be generated as .tsv files.)
}
\description{
Merge Results from MainSnvClass1.R
}
