% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainMergeSNVClass1.R
\name{MainMergeSNVClass1}
\alias{MainMergeSNVClass1}
\title{Merge Results from MainSNVClass1.R}
\usage{
MainMergeSNVClass1(hmdir = getwd(), annotation_file, input_dir, file_prefix)
}
\arguments{
\item{hmdir}{Home directory for the analysis (Default=getwd()).}

\item{annotation_file}{The result annotation file (${vcf}.${job_id}.peptide.txt) generated by MainSNVClass1(). 
For example, sample_vcf.txt.NO_job_id.peptide.txt.}

\item{input_dir}{(Required) Directory storing netMHCpan Results.}

\item{file_prefix}{(Required) File prefix of netMHCpan Results.
If you have "sample_annovar.txt.NO_JOB_ID.HLACLASS1.1.peptide.txt", please set "sample_annovar.txt.NO_JOB_ID".}
}
\value{
void (Calculated Neoantigen Files will be generated as .tsv files.):

HLA:  HLA type used to calculate neoantigen.

Pos:  The position of the fraction of peptide used to be evaluated from the full-length peptide.

Gene:  Gene symbol used to be evaluated in NetMHCpan.

Evaluated_Mutant_Peptide:  The mutant peptide to be evaluated.

Mut_IC50: IC50 value for evaluated mutant peptide.

Mut_Rank: Rank value for evaluated mutanat peptide.

Evaluated_Wt_Peptide: The wild-type peptide to be evaluated.

Wt_IC50: IC50 value for evaluated wild-type peptide.

Wt_Rank: Rank value for evaluated wild-type peptide.

Chr: Chromosome Number of the mutation.

NM_ID: NM_ID used to construct peptides from the mutation.

Change: The annotation to be described in .vcf file.

Ref: reference type nucleic acid base.

Alt: alternative type nucleic acid base.

Prob: A probability of reference nucleic acid base described in .vcf file.

Mutation_Prob: A probability of alternative nucleic acid base described in .vcf file.

Exon_Start: The exon start position of the corrsponding NM_ID.

Exon_End: The exon end position of the corrsponding NM_ID.

Mutation_Position: The mutation position of the corrsponding NM_ID.

Total_Depth: The depth of the reference nucleic acid base.

Tumor_Depth: The depth of the alternative nucleic acid base.

Wt_Peptide: The full-length of the wild-type peptide.

Mutant_Peptide: The full-length of the mutant peptide.

Total_RNA: The expression amount of the corresponding RNA.

Tumor_RNA_Ratio: The variant allele frequency of the corresponding RNA.

Tumor_RNA: The modified amount of the corresponding RNA level based on RNA Reads.

Tumor_RNA_based_on_DNA: The modified amount of the corresponding RNA level based on DNA Reads.

MutRatio: The mean value of the cancer cell fraction probability.

MutRatio_Min: The 1\% percentile of the cancer cell fraction probability.

MutRatio_Max: The 99\% percentile of the cancer cell fraction probability.
}
\description{
Merge Results from MainSNVClass1.R
}
