% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainSNVClass1.R
\name{MainSNVClass1}
\alias{MainSNVClass1}
\title{Calculate Neoantigen Candidates on SNVs for MHC Class1}
\usage{
MainSNVClass1(input_file, HLA_file, file_name_in_HLA_table = input_file,
  hmdir = getwd(), job_ID = "NO_JOB_ID", RNAseq_file = NA, RNA_bam = NA,
  CNV = NA, ccfp_dir = paste(hmdir, "lib/ccfp.jar", sep = ""),
  Purity = NA, netMHCpan_dir = paste(hmdir,
  "lib/netMHCIIpan-3.1/netMHCIIpan", sep = ""), refDNA = paste(hmdir,
  "lib/GRCh37.fa", sep = ""), refFlat_file = paste(hmdir, "/data/refFlat.txt",
  sep = ""), refMrna_1 = paste(hmdir, "/data/refMrna.cut1.fa", sep = ""),
  refMrna_3 = paste(hmdir, "/data/refMrna.cut3.fa", sep = ""),
  samtools_dir = "samtools", bcftools_dir = "bcftools", Chr_Column = 1,
  Mutation_Start_Column = 2, Mutation_End_Column = 3,
  Mutation_Ref_Column = 4, Mutation_Alt_Column = 5, NM_ID_Column = 10,
  Depth_Normal_Column = NA, Depth_Tumor_Column = NA,
  ambiguous_between_exon = 0, ambiguous_codon = 0, peptide_length = c(8,
  9, 10, 11, 12, 13))
}
\arguments{
\item{input_file}{An input file annotated by ANNOVAR (http://annovar.openbioinformatics.org/en/latest/) or other softwares to be analyzed (Required).
See by data(sample_annovar.txt); sample_annovar; data(sample_genomon); sample_genomon;}

\item{HLA_file}{A tab separated file indicating HLA types (Required).
The 1st column is input_file name, and the following columns indicate HLA types.
See by data(hla_table); hla_table;}

\item{file_name_in_HLA_table}{If the name (1st column) in HLA table is not input_file, indicate the corresponding name (Default=input_file).}

\item{hmdir}{Home directory for the analysis (Default=getwd()).}

\item{job_ID}{Job-Id to be attached in output files (Default="NO_JOB_ID").}

\item{RNAseq_file}{(Default=NA) A file including RNA expressions.
The 1st, 2nd and 3rd columns are "GeneSymbol Chr:ExonStart-ExonEnd(locus) Expression Amount", respectively.
The 1st row should be any header.
See by data(RNAseq.txt); RNAseq;}

\item{RNA_bam}{RNA bam file to calculate variant allele frequency of RNA at each mutation (Default=NA).}

\item{CNV}{A file including copy number variation to calculate cancer cell fraction probability (CCFP) (Default=NA).
The format is according to ASCAT (https://www.crick.ac.uk/peter-van-loo/software/ASCAT) output files.
The columns are "Chromosome Position Log R segmented LogR BAF segmented BAF Copy number Minor allele Raw copy number"
The 1st row should be the above header.
See data(CopyNum); head(CopyNum);}

\item{ccfp_dir}{The file directory to CCFP.pl (Default="lib/ccfp.jar").}

\item{Purity}{Tumor purity or tumor contents ratio required to calculate CCFP (Default=NA).}

\item{netMHCpan_dir}{The file directory to netMHCpan (Default="lib/netMHCpan-3.0/netMHCpan").}

\item{refDNA}{refDNA information to be used to calculate RNA VAF (Default="lib/GRCh37.fa").}

\item{refFlat_file}{refFlat file to be used in constructing peptide. (Default=paste(hmdir,"lib/refFlat.txt",sep="")
This file is automaticalluy generated by Main_Preparation.sh}

\item{refMrna_1}{refMrna file to be used in constructing peptide (Default=paste(hmdir,"lib/refMrna.merge.cut1.fa",sep="").
This file is automaticalluy generated by Main_Preparation.sh}

\item{refMrna_3}{refMrna file to be used in constructing peptide (Default=paste(hmdir,"lib/refMrna.merge.cut3.fa",sep="").
This file is automaticalluy generated by Main_Preparation.sh}

\item{Chr_Column}{The column number describing Chromosome number in input_file (Default=1).}

\item{Mutation_Start_Column}{The column number describing Mutation Start Position in input_file (Default=2) .}

\item{Mutation_End_Column}{The column number describing Mutation End Position in input_file (Default=3).}

\item{Mutation_Ref_Column}{The column number describing Mutation Ref in input_file (Default=4).}

\item{Mutation_Alt_Column}{The column number describing Mutation Alt in input_file (Default=5).}

\item{NM_ID_Column}{The column number describing NM IDs in input_file (Default=10).}

\item{Depth_Normal_Column}{The column number describing the read count from normal cells (Default = NA)}

\item{Depth_Tumor_Column}{The column number describing the read count from tumor cells (Default = NA)}

\item{ambiguous_between_exon}{The maximum number to permit the differences between Exon-Lengths from refFlat and refMrna (Default=0).}

\item{ambiguous_codon}{The maximum number to permit the differences between inputfile- and refMrna-oriented translation START/END position (Default=0).}

\item{peptide_length}{Peptide Length to be generated (Default={8,9,10,11,12,13}).}
}
\value{
void (Calculated Neoantigen Files will be generated as .tsv files.)
}
\description{
Calculate Neoantigen Candidates on SNVs for MHC Class1
}
