% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{package}
\name{Neoantimon}
\alias{Neoantimon}
\alias{Neoantimon-package}
\title{Neoantimon}
\description{
Calculate Lists of Candidate Neoantingens from SNVs, Indels, and SV fusions to MHC Class1 and Class2.
First use MainSNVClass1, MainSNVClass2, MainINDELClass1, MainINDELClass2, MainSVFUSIONClass1, and MainSVFUSIONClass2.
}
