% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample.snps.vcf}
\alias{sample.snps.vcf}
\title{A Format / Sample file for snp informatin.}
\format{
A data frame with variables including "#CHROM"	"POS"	"ID"	"REF"	"ALT"	"QUAL"	"FILTER"	"INFO"	"FORMAT".
}
\usage{
data(sample.snps.vcf)
}
\description{
A dataset containing snps information of a patient.
}
\keyword{datasets}
