% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_copynum}
\alias{sample_copynum}
\title{A Format / Sample file for Copy Number Information}
\format{A data frame with 7 rows and 9 variables}
\usage{
data(sample_copynum)
}
\description{
A dataset containing the copy number information obtained by, e.g., ASCAT.
}
\keyword{datasets}
