% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_hla_table_c1}
\alias{sample_hla_table_c1}
\title{A Format / Sample file for HLA CLASS1 Table}
\format{A data frame with 3 rows and at most 7 variables}
\usage{
data(sample_hla_table_c1)
}
\description{
A dataset containing the HLA types of patients in each row.
}
\keyword{datasets}
