% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_hla_table_c2}
\alias{sample_hla_table_c2}
\title{A Format / Sample file for HLA CLASS2 Table}
\format{
A data frame with at least 3 row and at most 10 variables
}
\usage{
data(sample_hla_table_c2)
}
\description{
A dataset containing the HLA types of patients in each row.
}
\keyword{datasets}
