% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_refFlat.grch37}
\alias{sample_refFlat.grch37}
\title{A Sample file for refFlat}
\format{A data frame with 11 column.}
\usage{
data(sample_refFlat.grch37)
}
\description{
A dataset containing a part of refFlat data.
}
\keyword{datasets}
