% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_refMrna.grch37.fa}
\alias{sample_refMrna.grch37.fa}
\title{A Sample file for refSeq RNA}
\format{
A data frame with 1 column.
}
\usage{
data(sample_refMrna.grch37.fa)
}
\description{
A dataset containing a part of refSeq RNA.
}
\keyword{datasets}
