% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_result_INDEL_CLASS2_ALL}
\alias{sample_result_INDEL_CLASS2_ALL}
\title{Analyzed Result for INDEL CLASS2}
\usage{
data(sample_result_INDEL_CLASS2_ALL)
}
\description{
Analyzed Result for INDEL CLASS2
}
\keyword{result}
