% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_result_SeqFragment_CLASS1_ALL}
\alias{sample_result_SeqFragment_CLASS1_ALL}
\title{Analyzed Result for A DNA Fragment CLASS1}
\usage{
data(sample_result_SeqFragment_CLASS1_ALL)
}
\description{
Analyzed Result for A DNA Fragment CLASS1
}
\keyword{result}
