% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_rna_exp}
\alias{sample_rna_exp}
\title{A Format / Sample file for RNA Expression Information}
\format{
A data frame with 22 rows and 3 variables
}
\usage{
data(sample_rna_exp)
}
\description{
A dataset containing the RNA expression amount of patient for each gene.
}
\keyword{datasets}
