% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_sv_bnd}
\alias{sample_sv_bnd}
\title{A Format / Sample file for Annotated vcf file.}
\format{
A data frame with 9 rows and variables including "Chr"	"Start"	"End"	"Ref"	"Alt (BND format)"	"Func.refGene (exonic, intron, intergenic, ...)"	"ExonicFunc.refGene (exonic nonsynonymous, synonymous, insertion, ...)"	"mateID (e.g., SVMERGE1_1)"
}
\usage{
data(sample_sv_bnd)
}
\description{
A dataset containing the variant information of a patient.
}
\keyword{datasets}
