% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_vcf}
\alias{sample_vcf}
\title{A Format / Sample file for Analyzed vcf file.}
\format{A data frame with 9 rows and 38 variables}
\usage{
data(sample_vcf)
}
\description{
A dataset containing the variant information of a patient.
}
\keyword{datasets}
