% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_vcf.snps}
\alias{sample_vcf.snps}
\title{A Format / Sample file for snp informatin.}
\format{A data frame with variables including "#CHROM"	"POS"	"ID"	"REF"	"ALT"	"QUAL"	"FILTER"	"INFO"	"FORMAT".}
\usage{
data(sample_vcf.snps)
}
\description{
A dataset containing snps information of a patient.
}
\keyword{datasets}
