% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoantimon.R
\docType{data}
\name{sample_vcf.vep}
\alias{sample_vcf.vep}
\title{A Format / Sample file for Annotated vcf file based on VEP.}
\format{
A data frame with variables including "#Uploaded_variation"	"Location"	"Allele"	"Gene"	"Feature"	"Feature_type"	"Consequence"	"cDNA_position"	"CDS_position"	"Protein_position"	"Amino_acids	Codons"	"Existing_variation"	"Extra"
}
\usage{
data(sample_vcf.vep)
}
\description{
A dataset containing the variant information of a patient.
}
\keyword{datasets}
