% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSampleFiles.R
\name{InstallSampleFiles}
\alias{InstallSampleFiles}
\title{Get Sample Files for Neoantimon}
\usage{

  InstallSampleFiles(url = "https://github.com/hase62/Neoantimon/raw/master/lib/data.zip",
  export_dir = "lib")
}
\arguments{
\item{url}{Url for getting samtools
(Default = "https://github.com/hase62/Neoantimon/raw/master/lib/data.zip").}

\item{export_dir}{Export directory (Default = "lib").}
}
\value{
void
}
\description{
Get Sample Files for Neoantimon
}
