% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainMergeSNVClass1.R
\name{MainMergeClass1}
\alias{MainMergeClass1}
\title{Merge Results from MainSnvClass1.R}
\usage{
MainMergeClass1(hmdir = getwd(), input_dir, input_file_prefix,
  Tumor_RNA_BASED_ON_DNA = TRUE, INDEL = FALSE)
}
\arguments{
\item{hmdir}{Home directory for the analysis (Default=getwd()).}

\item{input_dir}{Directory storing netMHCpan Results (Required).}

\item{input_file_prefix}{File prefix of netMHCpan Results (Required).
If you have "sample_annovar.txt.NO_JOB_ID.HLACLASS1.1.peptide.txt", please set "sample_annovar", "sample_annovar.txt" or "sample_annovar.txt.NO_JOB_ID".}

\item{Tumor_RNA_BASED_ON_DNA}{In calculating tumor specific RNA expression, TRUE uses variant allele frequency on DNA. Otherwise, use VAF on RNA (Default=TRUE).}

\item{INDEL}{If the targeting results are generated from Indels, Please check TRUE.}
}
\value{
void (Calculated Neoantigen Files will be generated as .tsv files.)

HLA A genoype of patient's HLAs used to be evaluated.

Pos The position of the fraction of peptide used to be evaluated from the full-length peptide.

Gene Gene symbol used to be evaluated in NetMHCpan.

MutatedPeptide The mutant peptide to be evaluated.

Mut_IC50 IC50 value for evaluated mutant peptide.

Mut_Rank Rank value for evaluated mutanat peptide.

Norm_Peptide The wild-type peptide to be evaluated.

Norm_IC50 IC50 value for evaluated wild-type peptide.

Norm_Rank Rank value for evaluated wild-type peptide.

Gene ID Gene symbol for the peptide.

Chr Chromosome Number of the mutation.

NM_ID NM_ID used to construct peptides from the mutation.

Change The annotation to be described in .vcf file.

ref reference type nucleic acid base.

alt alternative type nucleic acid base.

Prob A probability of reference nucleic acid base described in .vcf file.

MutationProb A probability of alternative nucleic acid base described in .vcf file.

ExonStart The exon start position of the corrsponding NM_ID.

ExonEnd The exon end position of the corrsponding NM_ID.

MutationPosition The mutation position of the corrsponding NM_ID.

Depth The depth of the reference nucleic acid base.

TumorDepth The depth of the alternative nucleic acid base.

PeptideNormal The full-length of the wild-type peptide.

PeptideMutation The full-length of the mutant peptide.

TotalRNA The expression amount of the corresponding RNA.

TumorRNARatio The variant allele frequency of the corresponding RNA.

TumorRNA The modified amount of the corresponding RNA level based on (RNA/DNA) VCF.

nA The total number of A allele copies.

nB The total number of B allele copies.

Checker CheckSum

MutRatio The mean value of the cancer cell fraction probability.

MutRatioMin The 1% percentile of the cancer cell fraction probability.

MutRatioMax The 99% percentile of the cancer cell fraction probability.
}
\description{
Merge Results from MainSnvClass1.R
}
